+++ helper/cambiarContrasenia.php
@@ -13,7 +13,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Cambio de password</title>
-    <link rel="stylesheet" href="../css/cambiar_contrasenia.css"> 
+    <link rel="stylesheet" href="../css/passwordCambiar.css"> 
 </head>
 <body>
     <div class="lineasup"></div>